requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
pydantic
python-dotenv
streamlit
plotly-express
sentence-transformers
python-multipart
scikit-learn
torch
unstructured[all-docs]
